latest sdk and 64 bit
Former-commit-id: 2a1b287d9f46da9d11e2cdf1ad447bc622a86db5
Former-commit-id: 2c582101b1ab79dab0865f625c79f019d2925d85

+++ shapes/ios/Classes/ShapeProxy.h
@@ -56,7 +56,7 @@ enum
 //@property(nonatomic,readonly) CALayer* layer;
 
 -(void)boundsChanged:(CGRect)bounds;
--(CGPoint) computePoint:(TiPoint*)center withAnchor:(int)anchor inSize:(CGSize)size decale:(CGSize)decale;
+-(CGPoint) computePoint:(TiPoint*)center withAnchor:(NSInteger)anchor inSize:(CGSize)size decale:(CGSize)decale;
 -(CGFloat*) arrayFromNSArray:(NSArray*)array;
 -(TiPoint *)defaultCenter;
 -(TiPoint *)defaultRadius;
